Blume_Chatbot/requirements.txt
Django==4.2.4
google-generativeai==0.1.0
gTTS==2.2.4
pygame==2.1.0
langdetect==1.0.9
PyPDF2==3.0.1
Pillow==9.3.0
mimetypes==0.1.3

